song-vocab/requirements.txt
ollama==0.1.6
instructor==0.4.5
pydantic==2.10.6
aiohttp==3.9.3
beautifulsoup4==4.12.3
python-dotenv
google-search-results==2.4.2
langdetect==1.0
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
gTTS==2.5.4
h11==0.14.0
idna==3.10
nest-asyncio==1.6.0
requests==2.32.3
urllib3==2.3.0
uvicorn==0.34.0
annotated-types==0.7.0
anyio==4.8.0
fastapi==0.115.10
pydantic_core==2.27.2
sniffio==1.3.1
starlette==0.46.0
typing_extensions==4.12.2